Refactor: RequestCard 컴포넌트 구조 개선 #87 - card-content의 data를 상위 컴포넌트에서 props로 넘기도록 개선
