+++ front-end/src/pages/student/course/request/StudentRequest.tsx
@@ -3,9 +3,49 @@ import RequestCard from './components/RequestCard';
 import S from './StudentRequest.module.css';
 import { classroomRepository } from '../../../../di';
 import SuccessPopup from '../components/SuccessPopup';
+import WishSvg from '../../../../assets/icons/wish-emoji.svg?react';
+import WindSvg from '../../../../assets/icons/wind-emoji.svg?react';
+import BulbSvg from '../../../../assets/icons/bulb-emoji.svg?react';
+import MagnifierSvg from '../../../../assets/icons/magnifier-emoji.svg?react';
+import EarSvg from '../../../../assets/icons/ear-emoji.svg?react';
+import {
+  RequestFast,
+  RequestHard,
+  RequestQuestion,
+  RequestSize,
+  RequestSound,
+} from '../../../../core/model';
 
 const CARD_TYPES = ['hard', 'fast', 'question', 'size', 'sound'] as const;
 export type CardType = 'hard' | 'fast' | 'question' | 'size' | 'sound';
+
+const CARD_CONTENT = {
+  hard: {
+    icon: <WishSvg className={S.icon} />,
+    title: RequestHard.title,
+    description: RequestHard.description,
+  },
+  fast: {
+    icon: <WindSvg className={S.icon} />,
+    title: RequestFast.title,
+    description: RequestFast.description,
+  },
+  question: {
+    icon: <BulbSvg className={S.icon} />,
+    title: RequestQuestion.title,
+    description: RequestQuestion.description,
+  },
+  size: {
+    icon: <MagnifierSvg className={S.icon} />,
+    title: RequestSize.title,
+    description: RequestSize.description,
+  },
+  sound: {
+    icon: <EarSvg className={S.icon} />,
+    title: RequestSound.title,
+    description: RequestSound.description,
+  },
+};
 
 const StudentRequest = ({ courseId }: { courseId: string }) => {
   const [successPopup, setSuccessPopup] = useState<boolean>(false);
@@ -33,8 +73,10 @@ const StudentRequest = ({ courseId }: { courseId: string }) => {
         {CARD_TYPES.map((type) => (
           <RequestCard
             key={type}
-            type={type}
             onCardClick={() => handleCardClick(type)}
+            title={CARD_CONTENT[type].title}
+            description={CARD_CONTENT[type].description}
+            icon={CARD_CONTENT[type].icon}
           />
         ))}
       </div>
