+++ front-end/src/pages/student/course/request/components/RequestCard.tsx
@@ -1,56 +1,21 @@
 import S from './RequestCard.module.css';
-import WishSvg from '../../../../../assets/icons/wish-emoji.svg?react';
-import WindSvg from '../../../../../assets/icons/wind-emoji.svg?react';
-import BulbSvg from '../../../../../assets/icons/bulb-emoji.svg?react';
-import MagnifierSvg from '../../../../../assets/icons/magnifier-emoji.svg?react';
-import EarSvg from '../../../../../assets/icons/ear-emoji.svg?react';
-import {
-  RequestFast,
-  RequestHard,
-  RequestQuestion,
-  RequestSize,
-  RequestSound,
-} from '../../../../../core/model';
-import { CardType } from '../StudentRequest';
 import useBlockTimer from '../../../../../hooks/useBlockTimer';
 import { useState } from 'react';
 
-const CARD_CONTENT = {
-  hard: {
-    icon: <WishSvg className={S.icon} />,
-    title: RequestHard.title,
-    description: RequestHard.description,
-  },
-  fast: {
-    icon: <WindSvg className={S.icon} />,
-    title: RequestFast.title,
-    description: RequestFast.description,
-  },
-  question: {
-    icon: <BulbSvg className={S.icon} />,
-    title: RequestQuestion.title,
-    description: RequestQuestion.description,
-  },
-  size: {
-    icon: <MagnifierSvg className={S.icon} />,
-    title: RequestSize.title,
-    description: RequestSize.description,
-  },
-  sound: {
-    icon: <EarSvg className={S.icon} />,
-    title: RequestSound.title,
-    description: RequestSound.description,
-  },
-};
-
 type RequestCardProps = {
-  type: CardType;
   onCardClick: () => Promise<boolean>;
+  icon: JSX.Element;
+  title: string;
+  description: string;
 };
 
-const RequestCard = ({ type, onCardClick }: RequestCardProps) => {
+const RequestCard = ({
+  onCardClick,
+  icon,
+  title,
+  description,
+}: RequestCardProps) => {
   const [isSelected, setIsSelected] = useState<boolean>(false);
-  const { icon, title, description } = CARD_CONTENT[type] || {};
   const { isBlocked, countdown } = useBlockTimer(
     isSelected,
     setIsSelected,
